requirements.txt
django-cms<3.3
django-treebeard>=2.0

https://github.com/divio/djangocms-column/archive/master.zip?1453417739.24
https://github.com/divio/djangocms-googlemap/archive/master.zip?1453417739.24
https://github.com/divio/djangocms-inherit/archive/master.zip?1453417739.24
https://github.com/divio/djangocms-link/archive/master.zip?1453417739.24
https://github.com/divio/djangocms-style/archive/master.zip?1453417739.24
https://github.com/divio/djangocms-file/archive/master.zip?1453417739.24
https://github.com/divio/djangocms-picture/archive/master.zip?1453417739.24
https://github.com/divio/djangocms-teaser/archive/master.zip?1453417739.24
https://github.com/divio/djangocms-video/archive/master.zip?1453417739.24
djangocms-admin-style>=1.0.6
djangocms-text-ckeditor>=2.8.0
Django<1.9
pytz
django-reversion>=1.8.7,<1.9
django-classy-tags>=0.3.4.1
html5lib
Pillow>=2.3
django-sekizai>=0.7
six